+++ packages/backend/src/core/StateUpdater.ts
@@ -1,13 +1,8 @@
-import {
-  ForcedAction,
-  OnChainData,
-  StarkExProgramOutput,
-} from '@explorer/encoding'
+import { ForcedAction, StarkExProgramOutput } from '@explorer/encoding'
 import { PositionLeaf, RollupState } from '@explorer/state'
 import { Hash256, PedersenHash, Timestamp } from '@explorer/types'
 
 import { ForcedTransactionsRepository } from '../peripherals/database/ForcedTransactionsRepository'
-import { PageRepository } from '../peripherals/database/PageRepository'
 import { PositionRecord } from '../peripherals/database/PositionRepository'
 import { RollupStateRepository } from '../peripherals/database/RollupStateRepository'
 import { StateTransitionRecord } from '../peripherals/database/StateTransitionRepository'
@@ -20,67 +15,17 @@ import { Logger } from '../tools/Logger'
  * @internal
  * Same as `await RollupState.empty().then(empty => empty.positions.hash())`
  */
-export const ROLLUP_STATE_EMPTY_HASH = PedersenHash(
-  '52ddcbdd431a044cf838a71d194248640210b316d7b1a568997ecad9dec9626'
-)
 
 export class StateUpdater {
   constructor(
-    private readonly pageRepository: PageRepository,
-    private readonly stateUpdateRepository: StateUpdateRepository,
-    private readonly rollupStateRepository: RollupStateRepository,
-    private readonly ethereumClient: EthereumClient,
-    private readonly forcedTransactionsRepository: ForcedTransactionsRepository,
-    private readonly logger: Logger,
-    private rollupState?: RollupState
+    protected readonly stateUpdateRepository: StateUpdateRepository,
+    protected readonly rollupStateRepository: RollupStateRepository,
+    protected readonly ethereumClient: EthereumClient,
+    protected readonly forcedTransactionsRepository: ForcedTransactionsRepository,
+    protected readonly logger: Logger,
+    protected readonly emptyStateHash: PedersenHash,
+    protected rollupState?: RollupState
   ) {}
-
-  async loadRequiredPages(
-    stateTransitions: Omit<StateTransitionRecord, 'id'>[]
-  ): Promise<(StateTransitionRecord & { pages: string[] })[]> {
-    if (stateTransitions.length === 0) {
-      return []
-    }
-
-    const pageGroups = await this.pageRepository.getByStateTransitions(
-      stateTransitions.map((x) => x.stateTransitionHash)
-    )
-    const stateTransitionsWithPages = pageGroups.map((pages, i) => {
-      const stateTransition = stateTransitions[i]
-      if (stateTransition === undefined) {
-        throw new Error('Programmer error: state transition count mismatch')
-      }
-      return { ...stateTransition, pages }
-    })
-    if (pageGroups.length !== stateTransitions.length) {
-      throw new Error('Missing pages for state transitions in database')
-    }
-
-    const { oldHash, id } = await this.readLastUpdate()
-    await this.ensureRollupState(oldHash)
-
-    return stateTransitionsWithPages.map((transition, i) => ({
-      id: id + i + 1,
-      ...transition,
-    }))
-  }
-
-  async processOnChainStateTransition(
-    stateTransitionRecord: StateTransitionRecord,
-    onChainData: OnChainData
-  ) {
-    if (!this.rollupState) {
-      throw new Error('Rollup state not initialized')
-    }
-    const newPositions = await this.rollupState.calculateUpdatedPositions(
-      onChainData
-    )
-    return this.processStateTransition(
-      stateTransitionRecord,
-      onChainData,
-      newPositions
-    )
-  }
 
   async processStateTransition(
     stateTransitionRecord: StateTransitionRecord,
@@ -158,12 +103,12 @@ export class StateUpdater {
     if (lastUpdate) {
       return { oldHash: lastUpdate.rootHash, id: lastUpdate.id }
     }
-    return { oldHash: ROLLUP_STATE_EMPTY_HASH, id: 0 }
+    return { oldHash: this.emptyStateHash, id: 0 }
   }
 
   async ensureRollupState(oldHash: PedersenHash, height?: bigint) {
     if (!this.rollupState) {
-      if (oldHash === ROLLUP_STATE_EMPTY_HASH) {
+      if (oldHash === this.emptyStateHash) {
         this.rollupState = await RollupState.empty(
           this.rollupStateRepository,
           height
