+++ packages/backend/test/core/StateUpdater.test.ts
@@ -1,26 +1,17 @@
-import { Block } from '@ethersproject/providers'
-import { MerkleTree, PositionLeaf, RollupState } from '@explorer/state'
-import {
-  AssetId,
-  Hash256,
-  PedersenHash,
-  StarkKey,
-  Timestamp,
-} from '@explorer/types'
+import { AssetId, Hash256, PedersenHash, StarkKey } from '@explorer/types'
 import { expect } from 'earljs'
 
-import {
-  ROLLUP_STATE_EMPTY_HASH,
-  StateUpdater,
-} from '../../src/core/StateUpdater'
+import { StateUpdater } from '../../src/core/StateUpdater'
 import { ForcedTransactionsRepository } from '../../src/peripherals/database/ForcedTransactionsRepository'
-import type { PageRepository } from '../../src/peripherals/database/PageRepository'
 import type { RollupStateRepository } from '../../src/peripherals/database/RollupStateRepository'
 import { StateUpdateRepository } from '../../src/peripherals/database/StateUpdateRepository'
 import type { EthereumClient } from '../../src/peripherals/ethereum/EthereumClient'
 import { Logger } from '../../src/tools/Logger'
-import { decodedFakePages } from '../fakes'
 import { mock } from '../mock'
+
+const EMPTY_STATE_HASH = PedersenHash(
+  '52ddcbdd431a044cf838a71d194248640210b316d7b1a568997ecad9dec9626'
+)
 
 describe(StateUpdater.name, () => {
   describe(StateUpdater.prototype.ensureRollupState.name, () => {
@@ -29,17 +20,17 @@ describe(StateUpdater.name, () => {
         persist: async () => {},
       })
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         mock<StateUpdateRepository>(),
         rollupStateRepository,
         mock<EthereumClient>(),
         mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
-      // ROLLUP_STATE_EMPTY_HASH is for tree of height 64 and  recalculating this hash
+      // EMPTY_STATE_HASH is for tree of height 64 and  recalculating this hash
       // in tests on slower machines (e.g. CI) makes test flakey async-wise.
       const rollupState = await stateUpdater.ensureRollupState(
-        ROLLUP_STATE_EMPTY_HASH,
+        EMPTY_STATE_HASH,
         3n
       )
       const rollupStateEmptyHashForHeight3 = PedersenHash(
@@ -53,12 +44,12 @@ describe(StateUpdater.name, () => {
 
     it('sets state after restart', async () => {
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         mock<StateUpdateRepository>(),
         mock<RollupStateRepository>(),
         mock<EthereumClient>(),
         mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
       const hash = PedersenHash.fake()
       const rollupState = await stateUpdater.ensureRollupState(hash)
@@ -67,12 +58,12 @@ describe(StateUpdater.name, () => {
 
     it('resets state after reorg', async () => {
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         mock<StateUpdateRepository>(),
         mock<RollupStateRepository>(),
         mock<EthereumClient>(),
         mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
       const hashA = PedersenHash.fake('a')
       const hashB = PedersenHash.fake('b')
@@ -84,122 +75,17 @@ describe(StateUpdater.name, () => {
 
     it('leaves state intact before a subsequent state transition', async () => {
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         mock<StateUpdateRepository>(),
         mock<RollupStateRepository>(),
         mock<EthereumClient>(),
         mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
       const hash = PedersenHash.fake()
       const rollupStateA = await stateUpdater.ensureRollupState(hash)
       const rollupStateB = await stateUpdater.ensureRollupState(hash)
       expect(rollupStateA).toReferentiallyEqual(rollupStateB)
-    })
-  })
-
-  describe(StateUpdater.prototype.loadRequiredPages.name, () => {
-    it('throws if pages are missing in database', async () => {
-      const pageRepository = mock<PageRepository>({
-        getByStateTransitions: async () => [],
-      })
-      const stateUpdater = new StateUpdater(
-        pageRepository,
-        mock<StateUpdateRepository>(),
-        mock<RollupStateRepository>(),
-        mock<EthereumClient>(),
-        mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
-      )
-      await expect(
-        stateUpdater.loadRequiredPages([
-          { stateTransitionHash: Hash256.fake('a'), blockNumber: 1 },
-        ])
-      ).toBeRejected('Missing pages for state transitions in database')
-    })
-
-    it('returns correct StateTransition for every update', async () => {
-      const pageRepository = mock<PageRepository>({
-        getByStateTransitions: async () => [
-          ['aa', 'ab', 'ac'],
-          ['ba', 'bb'],
-        ],
-      })
-      const stateUpdateRepository = mock<StateUpdateRepository>({
-        findLast: async () => ({
-          rootHash: PedersenHash.fake('1234'),
-          id: 567,
-          timestamp: Timestamp(1),
-          blockNumber: Math.random(),
-          stateTransitionHash: Hash256.fake(),
-        }),
-      })
-      const stateUpdater = new StateUpdater(
-        pageRepository,
-        stateUpdateRepository,
-        mock<RollupStateRepository>(),
-        mock<EthereumClient>(),
-        mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
-      )
-
-      const stateTransitions = await stateUpdater.loadRequiredPages([
-        { blockNumber: 123, stateTransitionHash: Hash256.fake('123') },
-        { blockNumber: 456, stateTransitionHash: Hash256.fake('456') },
-      ])
-      expect(stateTransitions).toEqual([
-        {
-          id: 567 + 1,
-          blockNumber: 123,
-          stateTransitionHash: Hash256.fake('123'),
-          pages: ['aa', 'ab', 'ac'],
-        },
-        {
-          id: 567 + 2,
-          blockNumber: 456,
-          stateTransitionHash: Hash256.fake('456'),
-          pages: ['ba', 'bb'],
-        },
-      ])
-    })
-  })
-
-  describe(StateUpdater.prototype.processOnChainStateTransition.name, () => {
-    it('throws if calculated root hash does not match the one from verifier', async () => {
-      const collector = new StateUpdater(
-        mock<PageRepository>(),
-        mock<StateUpdateRepository>(),
-        mock<RollupStateRepository>(),
-        mock<EthereumClient>({
-          getBlock: async () => {
-            return { timestamp: 1 } as unknown as Block
-          },
-        }),
-        mock<ForcedTransactionsRepository>(),
-        Logger.SILENT,
-        mock<RollupState>({
-          calculateUpdatedPositions: async () => [
-            { index: 1n, value: mock<PositionLeaf>() },
-          ],
-          update: async () =>
-            ({
-              positionTree: mock<MerkleTree<PositionLeaf>>({
-                hash: async () => PedersenHash.fake('1234'),
-              }),
-            } as unknown as RollupState),
-        })
-      )
-
-      await expect(
-        collector.processOnChainStateTransition(
-          {
-            id: 1,
-            stateTransitionHash: Hash256.fake('123'),
-            blockNumber: 1,
-          },
-          decodedFakePages
-        )
-      ).toBeRejected('State transition calculated incorrectly')
     })
   })
 
@@ -210,12 +96,12 @@ describe(StateUpdater.name, () => {
       })
 
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         stateUpdateRepository,
         mock<RollupStateRepository>(),
         mock<EthereumClient>(),
         mock<ForcedTransactionsRepository>(),
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
 
       await stateUpdater.discardAfter(20)
@@ -235,12 +121,12 @@ describe(StateUpdater.name, () => {
       })
 
       const stateUpdater = new StateUpdater(
-        mock<PageRepository>(),
         mock<StateUpdateRepository>(),
         mock<RollupStateRepository>(),
         mock<EthereumClient>(),
         forcedTransactionsRepository,
-        Logger.SILENT
+        Logger.SILENT,
+        EMPTY_STATE_HASH
       )
 
       await expect(
